+++ CHANGELOG.md
@@ -5,6 +5,9 @@ From v3.0.1 to v3.1.0
 ## Fixes
 
 - Fixed a bug in `QSPRDataset` where property transformations were not applied.
+- Fixed a bug where an attached standardizer would be refit when calling 
+`QSPRModel.predictMols` with `use_applicability_domain=True`.
+- Fixed random seed not set in `FoldsFromDataSplit.iterFolds` for `ClusterSplit`.
 
 ## Changes
 
@@ -33,3 +36,5 @@ From v3.0.1 to v3.1.0
       cleared from the data set with the `clear` parameter of `dropDescriptorSets`.
 
 ## Removed Features
+
+None.